added routing for ChooseTemplate, ListTemplate, and User
