+++ one2Do/one2Do/Controllers/CategoriesController.cs
@@ -11,7 +11,7 @@ namespace one2Do;
 
 public class CategoriesController : Controller
 {   
-    private one2doDbContext context; //
+    public one2doDbContext context; 
 
     public CategoriesController(one2doDbContext dbContext)
     {
@@ -26,10 +26,4 @@ public class CategoriesController : Controller
         return View();
     }
 
-//     [HttpGet]
-
-//     public IActionResult ////////do i need this? See codingEvents line 39 in EventCategoryController.cs
-//     //this is checkign to see if model is valid....
-
  }
-
