--- one2Do/one2Do/Controllers/ChooseTemplate.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace one2Do;
-
-public class ChooseTemplateController: Controller
-{
-    private readonly one2doDbContext context;
-    public ChooseTemplateController(one2doDbContext dbContext)
-    {
-        context=dbContext;
-    }
-    public IActionResult Index()
-    {
-        return View();
-    }
-}
