+++ one2Do/one2Do/Views/ChooseTemplate/Index.cshtml
@@ -10,3 +10,14 @@
         Here's some imaginary templates to choose from.
     </p>
 </div>
+
+<div>
+    <form method = "get">
+                <button asp-controller="user" asp-action="index" class="btn btn-primary"> Back to Profile</button>
+            </form>   
+
+            
+    <form method = "get">
+                <button asp-controller="ListTemplate" asp-action="index" class="btn btn-primary"> Choose Template to Edit</button>
+            </form>   
+</div>
