+++ one2Do/one2Do/Views/ListTemplate/Index.cshtml
@@ -10,3 +10,12 @@
         This is a list
     </p>
 </div>
+ <form method = "get">
+                <button asp-controller="user" asp-action="index" class="btn btn-primary"> Back to Profile</button>
+            </form>   
+
+            
+    <form method = "get">
+                <button asp-controller="user" asp-action="Todo
+                " class="btn btn-primary"> Save</button>
+            </form>
